+++ hadoop-hdfs-project/hadoop-hdfs/src/test/java/se/sics/hop/erasure_coding/TestUtil.java
@@ -11,7 +11,7 @@ public class TestUtil {
 
   public static void createRandomFile(DistributedFileSystem dfs, Path path, long seed, int blockCount,
         int blockSize) throws IOException {
-    FSDataOutputStream out = dfs.create(path, new EncodingPolicy("src", 1));
+    FSDataOutputStream out = dfs.create(path, new EncodingPolicy("src", (short) 1));
     byte[] buffer = randomBytes(seed, blockCount, blockSize);
     out.write(buffer, 0, buffer.length);
     out.close();
